change points type to double because Season: 2009 Round: 5 RedBull: 38.5 points [https://ergast.com/api/f1/2009/5/constructorStandings](https://ergast.com/api/f1/2009/5/constructorStandings)

+++ ergast/src/main/java/ru/olerom/formula/ergast/objects/ConstructorStandings.java
@@ -8,11 +8,11 @@ package ru.olerom.formula.ergast.objects;
 public class ConstructorStandings {
     private int position;
     private String positionText;
-    private int points;
+    private double points;
     private int wins;
     private Constructor constructor;
 
-    public ConstructorStandings(int position, String positionText, int points, int wins, Constructor constructor) {
+    public ConstructorStandings(int position, String positionText, double points, int wins, Constructor constructor) {
         this.position = position;
         this.positionText = positionText;
         this.points = points;
@@ -28,7 +28,7 @@ public class ConstructorStandings {
         return positionText;
     }
 
-    public int getPoints() {
+    public double getPoints() {
         return points;
     }
